# swagger.JSON
{
  "openapi": "3.0.0",
  "info": {
    "title": "User Auth API",
    "description": "API para autenticação de usuários — registro, confirmação, login, password reset e perfil.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:{port}",
      "description": "Servidor local",
      "variables": { "port": { "default": "3000" } }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserRegister": {
        "type": "object",
        "required": ["email","password"],
        "properties": {
          "email": { "type":"string","format":"email","example":"user@example.com" },
          "password": { "type":"string","minLength":6,"example":"senha123" }
        }
      },
      "UserLogin": {
        "type": "object",
        "required": ["email","password"],
        "properties": {
          "email": { "type":"string","format":"email","example":"user@example.com" },
          "password": { "type":"string","example":"senha123" }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type":"string","example":"607f1f77bcf86cd799439011" },
          "email": { "type":"string","format":"email","example":"user@example.com" },
          "createdAt": { "type":"string","format":"date-time" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type":"string","example":"Descrição do erro" }
        }
      }
    }
  },
...
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Registrar novo usuário",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserRegister" } } } },
        "responses": {
          "201": { "description": "Usuário criado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "400": { "description": "Erro de validação", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Fazer login e obter token JWT",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserLogin" } } } },
        "responses": {
          "200": {
            "description": "Login bem-sucedido",
            "headers": {
              "Authorization": { "schema": { "type":"string","example":"Bearer <token>" } }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type":"object",
                  "properties": {
                    "token": { "type":"string","example":"<token>" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "description": "Credenciais inválidas", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}